bugfix
git-svn-id: file:///Users/braun/svn/vermont/branches/vermont-dynamic-config@1984 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ common/AgeBloomFilter.cpp
@@ -18,7 +18,7 @@
 
 #include "AgeBloomFilter.h"
 
-void AgeArray::resize(uint32_t size)
+void AgeArray::resize(size_t size)
 {
 	free(array);
 	array_size = size;
